CHANGE: 전화번호, 생일 빈값 처리, thread.dao 파일 병합, 오탈자 처리
