+++ routes/index.js
@@ -3,8 +3,6 @@ const express = require("express");
 const { userRouter } = require("./user.route");
 const { threadRouter } = require("./thread.route");
 
-const { threadRouter } = require("./thread.route");
-
 const routes = express.Router();
 
 routes.use("/user", userRouter);
